requirements.txt
pandas
python-dotenv
requests
google-api-python-client
google-auth
google-auth-httplib2
matplotlib
gspread
apscheduler
python-telegram-bot==20.7
clickhouse-connect



